[FEAT] : cage 외곽 점 구하는 로직 구현 [STYLE] : cage 스킬 사용 위치 점 9개 -> 25개로 변경

+++ Project/Assets/03.Code/Scripts/RGLightGame/Player.cs
@@ -5,58 +5,58 @@ using UnityEngine;
 
 namespace Minigame.RGLight
 {
-    public class Player : MonoBehaviour
-    {
-        public float moveSpeed;
-        public int maxHealth;
-        private int _curHealth;
+	public class Player : MonoBehaviour
+	{
+		public float moveSpeed;
+		public float maxHealth;
+		private float _curHealth;
 
-        public bool isDead;
+		public bool isDead;
 
-        public RGLightManager RGLightManager { get; private set; }
+		public RGLightManager RGLightManager { get; private set; }
 
-        public PlayerDistanceTracker PlayerDistanceTracker { get; private set; }
-        public PlayerRay PlayerRay { get; private set; }
-        private PlayerInputManager _playerInputManager;
-        private Rigidbody _playerRigidbody;
+		public PlayerDistanceTracker PlayerDistanceTracker { get; private set; }
+		public PlayerRay PlayerRay { get; private set; }
+		private PlayerInputManager _playerInputManager;
+		private Rigidbody _playerRigidbody;
 
-        private void FixedUpdate()
-        {
-            if (RGLightManager.IsEndGame) return;
-            Move(_playerInputManager.InputMoveDir);
-        }
+		private void FixedUpdate()
+		{
+			if (RGLightManager.IsEndGame) return;
+			Move(_playerInputManager.InputMoveDir);
+		}
 
-        private void Move(Vector2 input)
-        {
-            Vector3 actualMove = new Vector3(input.x, 0, input.y).normalized * moveSpeed * Time.fixedDeltaTime;
-            _playerRigidbody.MovePosition(_playerRigidbody.position + actualMove);
-        }
+		private void Move(Vector2 input)
+		{
+			Vector3 actualMove = new Vector3(input.x, 0, input.y).normalized * moveSpeed * Time.fixedDeltaTime;
+			_playerRigidbody.MovePosition(_playerRigidbody.position + actualMove);
+		}
 
-        public void TakeDamage(int damage)
-        {
-            _curHealth -= damage;
-            if (_curHealth <= 0 && !isDead) Die();
-        }
+		public void TakeDamage(float damage)
+		{
+			_curHealth -= damage;
+			if (_curHealth <= 0 && !isDead) Die();
+		}
 
-        private void Die()
-        {
-            isDead = true;
-            _curHealth = 0;
-            RGLightManager.GameResult(false);
-        }
+		private void Die()
+		{
+			isDead = true;
+			_curHealth = 0;
+			RGLightManager.GameResult(false);
+		}
 
-        public void Init(RGLightManager manager)
-        {
-            _playerInputManager = GetComponent<PlayerInputManager>();
-            _playerRigidbody = GetComponent<Rigidbody>();
-            PlayerDistanceTracker = GetComponent<PlayerDistanceTracker>();
-            PlayerRay = GetComponent<PlayerRay>();
-            RGLightManager = manager;
-            CinemachineVirtualCamera cvc = GameObject.FindObjectOfType<CinemachineVirtualCamera>();
-            cvc.Follow = transform;
-            cvc.LookAt = transform;
+		public void Init(RGLightManager manager)
+		{
+			_playerInputManager = GetComponent<PlayerInputManager>();
+			_playerRigidbody = GetComponent<Rigidbody>();
+			PlayerDistanceTracker = GetComponent<PlayerDistanceTracker>();
+			PlayerRay = GetComponent<PlayerRay>();
+			RGLightManager = manager;
+			CinemachineVirtualCamera cvc = GameObject.FindObjectOfType<CinemachineVirtualCamera>();
+			cvc.Follow = transform;
+			cvc.LookAt = transform;
 
-            _curHealth = maxHealth;
-        }
-    }
+			_curHealth = maxHealth;
+		}
+	}
 }
